requirements.txt
aiohttp
requests
Pillow
Telethon